+++ extensions/tidyread---streamline-your-daily-reading/src/add-source.command.tsx
@@ -1,5 +1,10 @@
 import RSSSearch from "./components/RSSSearch";
+import RedirectRoute from "./components/RedirectRoute";
 
 export default function CreateSourceForm() {
-  return <RSSSearch />;
+  return (
+    <RedirectRoute>
+      <RSSSearch />
+    </RedirectRoute>
+  );
 }