+++ extensions/tidyread---streamline-your-daily-reading/src/components/GenTodaysDigestPanel.tsx
@@ -79,14 +79,16 @@ export default function GenTodaysDigestPanel({
   const [failedItemsNum, setFailedItemsNum] = useState(0);
   const [rawItemsNum, setRawItemsNum] = useState(0);
 
-  const md = `
+  const middleStageStr =
+    status === "success"
+      ? ""
+      : `
   ${
-    status === "failed"
-      ? `> ❗**Digest failed to generate**, error is: \`${errorMessage}\`. View related [doc](https://tidyread.info/docs/why-digest-failed) to know more.\n`
+    translateStatus !== "waiting"
+      ? `
+  🔄 \`Translate Titles\`  ${translateStatusMap[translateStatus]}\n`
       : ""
   }
-  📊 \`Total Items\`  ${pullItemsStatus === "success" ? `**${total}**` : pullItemsStatusMap[pullItemsStatus]}\n
-  🔄 \`Translate Titles\`  ${translateStatusMap[translateStatus]}\n
   ${
     sumarizeItemStatus === "start"
       ? `
@@ -96,16 +98,29 @@ export default function GenTodaysDigestPanel({
   `
       : ""
   }\n
+  `;
+
+  const md = `
+  ${
+    status === "failed"
+      ? `> ❗**Digest failed to generate**, error is: \`${errorMessage}\`. View related [doc](https://tidyread.info/docs/why-digest-failed) to know more.\n`
+      : ""
+  }
+  📊 \`Total Items\`  ${pullItemsStatus === "success" ? `**${total}**` : pullItemsStatusMap[pullItemsStatus]}\n
+  ${middleStageStr}
   ⌛ \`Total Time\`  ${totalTime ? `**${formatSeconds(totalTime)}**` : "--"}
-  
+
+  ${status === "success" ? "![digest_complete](./digest_done.svg)" : ""}
+
   ### 💡 Tips
   - Can't stand manually generating? Check [this](https://tidyread.info/docs/automate-daily-digest) to free your hand.
   - Generating process is too slow? Check [this](https://tidyread.info/docs/why-digest-failed#excessive-execution-time) to speed up.
+  - Looking for more reading sources? Check [this](https://tidyread.info/docs/where-to-find-rss) to solve it.
   `;
 
   const manageSourceListActionNode = (
     <Action
-      title="Manage Source List"
+      title="Manage Sources"
       shortcut={Keyboard.Shortcut.Common.Edit}
       icon={Icon.Pencil}
       onAction={() => {
@@ -200,6 +215,22 @@ export default function GenTodaysDigestPanel({
     }
   };
 
+  const retry = async () => {
+    setStatus("generating");
+    setPullItemsStatus("waiting");
+    setTranslateStatus("waiting");
+    setSumarizeItemStatus("waiting");
+    setDigest(null);
+    setErrorMessage("");
+    setTotal(0);
+    setTotalTime(0);
+    setSuccessItemsNum(0);
+    setFailedItemsNum(0);
+    setRawItemsNum(0);
+
+    await handleGenDigest();
+  };
+
   useEffect(() => {
     handleGenDigest();
   }, []);
@@ -218,6 +249,7 @@ export default function GenTodaysDigestPanel({
               }}
             />
           )}
+          {status === "failed" && <Action title="Retry" onAction={retry}></Action>}
           {status !== "generating" && (
             <>
               <Action.CopyToClipboard title="Copy Content" content={md} />