+++ extensions/tidyread---streamline-your-daily-reading/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Add Search RSS Page] - 2024-01-30
 - add search rss page
+- add redirect route to improve onboarding experience
+- improve gen digest panel
 - fix default maxItemsPerFeed 10 => 5